+++ modules/Maze/MazeMap.cc
@@ -10,8 +10,8 @@ MazeMap::MazeMap() {
 MazeMap::~MazeMap() {}
 
 void MazeMap::UpdateMazeMap() {
-    for (size_t i = 0; i < size.y; i++) {
-        for (size_t j = 0; j < size.x; j++) {
+    for (int i = 0; i < size.y; i++) {
+        for (int j = 0; j < size.x; j++) {
             if (map[i][j] != MAP_EDGE && map[i][j] != MAP_OUT) {
                 SetElem(vec_2i(j, i), MAP_EMPTY);
                 SetTarget(vec_2i(j, i), MAP_EMPTY);
